requirements.txt
anthropic>=0.7.0
mistralai>=0.0.7
groq>=0.3.0
ollama>=0.1.5
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
python-dotenv==1.0.0
sqlalchemy==2.0.23
aiohttp==3.9.1
prometheus-client==0.21.0
redis>=5.0.1
celery>=5.3.6
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
requests>=2.31.0
pandas>=2.1.4
numpy>=1.26.3
scikit-learn>=1.3.2
httpx==0.25.2
alembic==1.13.0
psycopg2-binary==2.9.9
